Pivot windows CI to github actions
Currently windows CI is configured to run in appveyor however appveyor
is not configured to run on rust-numpy in the PyO3 github organization
so there has been no CI running on windows. This commit fixes this by
switching the windows CI to run in github actions instead of using
appveyor. This makes it consistent with how windows testing is run with
PyO3/PyO3.

Fixes #128