Remove unsound custom element example.
While we do check the array data element during extract, we do not check the
type of the individual elements and hence there is no guarantee that they are of
type T or even of uniform type.

But if we did check each element, Python code having a reference the (base)
array could change the elements and thereby their types after our check leading
to type confusion in safe Rust code.