Small buffer optimization for InvertedAxes
Since InvertedAxes is basically a set of small integers, we can avoid dynamic
allocations by using a word-sized bitmap in the common case of up to 32 axes.

This should also not unduly increase code slice as there is no dynamic spilling
into the large representation, but the decision is rather taken once based on
the dimensionality of the array.