Drop bespoke GILOnceCell implementation used to cache capsule API
While we do not strictly need all the capabilities and guarantees of PyO3's
GILOnceCell, reducing the amount of open-coded synchronization seems like it is
worth it to me. It also better centralizes GIL-based synchronization concerns in
the PyO3 crate where alternative implementations better matching the Python
implementation could be provided.