Exclude some unnecessary files from the crate
Exclude from the published crate the contents of `.github/` (CI
workflows, etc.) and other files that are only useful for upstream
development, not for crate users (`.gitignore`, `codecov.yml`, `x.py`).
This slightly slims down the crate.

This is also helpful downstream, e.g. in Fedora, where a proposed
`rust-numpy-devel` package would have an automatically-generated
dependency on `/usr/bin/python3` if `x.py` is not excluded.

For now, tests and benchmarks are both still included in the published
crate.